14. Przeniesienie klasy Node do pliku IntLinkedList w celu pozbycia sie klasy która nie musi byc klasą publiczną.
